dotnotes/requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.0.5
djangorestframework==3.14.0
pytz==2022.7.1
sqlparse==0.4.2
